Remove name member attribute from Teacher and Student classes (never used in algorithm) Added all current test files to CMakeList and Makefile files Removed unused includes Removed dead code previously used for performances measurements in php_pro_web_launcher file Rename test folder to tests Now using g++ with -static lDFLAGS only for fully static build

+++ src/Room.cpp
@@ -2,6 +2,5 @@
 
 #include <utility>
 
-#include "config.h"
 
-Room::Room(const int pId, string pTag) : mId(pId), mTag(std::move(pTag)) {}
+Room::Room(const unsigned short int pId, string pTag) : mId(pId), mTag(std::move(pTag)) {}
